requirements.txt
future==1.0.0
iso8601==2.1.0
MouseInfo==0.1.3
numpy==2.3.3
pillow==11.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
pyopengltk==0.0.4
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
PyYAML==6.0.3
screeninfo==0.8.1
serial==0.0.97
six==1.17.0
